Indexing codebases and querying indexes (#1)
This patch provides a very bare-bones indexing and querying functionality. Queries are written in regular Erlang syntax and variables are unified against the indexed forms.

That is, running a shell on this project and trying the following:

```erlang
> glass:index(glass, ".").
> glass:search(glass, "maps:without([position], _)").
```

Would find two instances where a similar AST unifies:

```
%% in glass_to_query/1 at ./_build/default/lib/glass_backend/ebin/glass_query.beam:25
30| maps:without([position], Attrs)

%% in minimal_attributes/1 at ./_build/default/lib/glass_backend/ebin/glass_query.beam:90
91| maps:without([position], Map)
```

Variables follow similar semantics as Erlang's unification algorithm, but unification is performed in a strictly left-to-right fashion. Fresh bindings always succeed. Bound unifications succeed when the bound value is the same AST subtree.

Current caveats: because the internal format Glass uses is not the same as the Erlang's AST format, not all forms are supported yet (as they lack the transformation to/from the internal  format). It's also only possible to write queries on expressions currently.